refactor: Money 티켓매수계산 역할 LottoStore에 양도
