+++ src/main/java/lotto/domain/Money.java
@@ -17,7 +17,7 @@ public class Money {
         }
     }
 
-    public int getAffordableLottoTickets() {
-        return money / LottoStore.LOTTO_PRICE;
+    public int getMoney() {
+        return money;
     }
 }
